+++ src/main/java/com/example/jomajomadelivery/user/entity/User.java
@@ -1,7 +1,7 @@
 package com.example.jomajomadelivery.user.entity;
 
 import com.example.jomajomadelivery.address.entity.Address;
-import com.example.jomajomadelivery.auth.oauth.SocialProvider;
+import com.example.jomajomadelivery.auth.oauth2.SocialProvider;
 import com.example.jomajomadelivery.common.BaseEntity;
 import com.example.jomajomadelivery.user.dto.request.SignUpUserDto;
 import com.example.jomajomadelivery.user.dto.request.UserUpdateDto;
